Introduce genesis_config.rs
