+++ examples/demo-stf/src/app.rs
@@ -1,6 +1,6 @@
-#[cfg(feature = "native")]
-use crate::config::Config;
 use crate::runtime::Runtime;
+#[cfg(feature = "native")]
+use crate::stf_config::Config;
 use crate::tx_hooks_impl::DemoAppTxHooks;
 use crate::tx_verifier_impl::DemoAppTxVerifier;
 use sov_app_template::AppTemplate;
