+++ examples/demo-stf/src/tests/stf_tests.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 pub mod test {
     use crate::{
-        config::{create_demo_config, LOCKED_AMOUNT, TEST_SEQUENCER_DA_ADDRESS},
+        genesis_config::{create_demo_config, LOCKED_AMOUNT, TEST_SEQUENCER_DA_ADDRESS},
         runtime::Runtime,
         tests::{create_new_demo, data_generation::simulate_da, new_test_blob, C},
     };
